fukery

+++ calc/ftc_eval.py
@@ -3,6 +3,6 @@
 if __name__ == '__main__':
     from math import *
     exec("f = lambda x: {}".format(input("Input your function: lambda x: ")))
-    a = int(input("Lower limit: "))
-    b = int(input("Upper limit: "))
+    a = float(input("Lower limit: "))
+    b = float(input("Upper limit: "))
     print("Result: {}".format(f(b) - f(a)))
